Ids based on a number of adjectives and an animal

Ids based on a number of adjectives and an animal

adjective_animal(
  n = 1,
  n_adjectives = 1,
  style = "snake",
  max_len = Inf,
  alliterate = FALSE
)

Arguments

n
number of ids to return. If NULL, it instead returns the generating function

n_adjectives
Number of adjectives to prefix the animal with

style
Style to join words with. Can be one of "Pascal", "camel", "snake", "kebab", "dot", "title", "sentence", "lower", "upper", "constant" or "spongemock"

max_len
The maximum length of a word part to include (this may be useful because some of the names are rather long. This stops you generating a hexakosioihexekontahexaphobic_queenalexandrasbirdwingbutterfly). A vector of length 2 can be passed in here in which case the first element will apply to the adjectives (all of them) and the second element will apply to the animals.

alliterate
Produce "alliterative" adjective animals (e.g., hessian_hamster). Note that this cannot provide an equal probability of any particular combination because it forces a weighted sampling. Adjectives may also be repeated if n_adjectives is more than 1.

Details

The list of adjectives and animals comes from https://github.com/a-type/adjective-adjective-animal, and in turn from <gfycat.com>

Author

Rich FitzJohn

Examples

# Generate a random identifier:
ids::adjective_animal()
#> [1] "nonspheric_siberiantiger"

# Generate a bunch all at once:
ids::adjective_animal(5)
#> [1] "sacred_pachyderm"        "artsycraftsy_astarte"   
#> [3] "eccentric_doe"           "compassionless_drever"  
#> [5] "unfeared_americancicada"

# Control the style of punctuation with the style argument:
ids::adjective_animal(style = "lower")
#> [1] "ecological snake"
ids::adjective_animal(style = "CONSTANT")
#> [1] "FROSTED_DUNGBEETLE"
ids::adjective_animal(style = "camel")
#> [1] "deferentialCopepod"
ids::adjective_animal(style = "kebab")
#> [1] "frantic-grub"
ids::adjective_animal(style = "spongemock")
#> [1] "RAGiNg-lLaMa"

# Control the number of adjectives used
ids::adjective_animal(n_adjectives = 3)
#> [1] "intercontradictory_semimonarchical_mirthful_toucan"

# This can get out of hand quickly though:
ids::adjective_animal(n_adjectives = 7)
#> [1] "libertarian_parochialist_rash_engaged_evileyed_crystallitic_unangry_krill"

# Limit the length of adjectives and animals used:
ids::adjective_animal(10, max_len = 6)
#>  [1] "turbid_godwit" "bored_eyas"    "mutual_titi"   "hilly_anole"  
#>  [5] "fast_swift"    "lead_bird"     "glassy_cow"    "lax_xeme"     
#>  [9] "nonfat_liger"  "rainy_walrus" 

# The lengths can be controlled for adjectives and animals
# separately, with Inf meaning no limit:
ids::adjective_animal(10, max_len = c(6, Inf), n_adjectives = 2)
#>  [1] "webby_mature_ballpython"         "legal_terse_dromaeosaur"        
#>  [3] "mucky_fierce_northernspottedowl" "acrid_fussy_marmoset"           
#>  [5] "public_faulty_australiancurlew"  "wroth_mid_chimpanzee"           
#>  [7] "golden_united_nuthatch"          "poetic_wary_abyssiniancat"      
#>  [9] "cream_pebbly_civet"              "chaste_naggy_hippopotamus"      

# Pass n = NULL to bind arguments to a function
id <- ids::adjective_animal(NULL, n_adjectives = 2,
                            style = "dot", max_len = 6)
id()
#> [1] "bardic.green.hake"
id(10)
#>  [1] "united.cruel.barb"    "dreary.duckie.vixen"  "brashy.ambery.eft"   
#>  [4] "leafed.almond.tick"   "cosmic.lethal.turkey" "even.bleary.possum"  
#>  [7] "sudden.fleecy.bovine" "intact.tan.moray"     "chatty.front.urchin" 
#> [10] "tidal.maize.urva"    

# Alliterated adjective animals always aid added awesomeness
ids::adjective_animal(10, n_adjectives = 3, alliterate = TRUE)
#>  [1] "streetsmart_sour_stoned_seahorse"                          
#>  [2] "agitable_apidological_antidotical_altiplanochinchillamouse"
#>  [3] "sulfuryl_satisfactory_suboceanic_siamesecat"               
#>  [4] "degrading_dermographic_deceased_dormouse"                  
#>  [5] "selfsatisfied_scholarless_super_serval"                    
#>  [6] "tsarist_tame_termitic_turkey"                              
#>  [7] "censorious_creatural_camoflage_cormorant"                  
#>  [8] "adobe_aweinspiring_aware_anhinga"                          
#>  [9] "phantasmagoric_pseudoliberal_pseudohistoric_pharaohhound"  
#> [10] "adventurous_apian_atrocious_arctichare"