Random hexadecimal identifiers

Random hexadecimal identifiers. If possible, by default this uses the openssl package to produce a random set of bytes, and expresses that as a hex character string, creating cryptographically secure (unpredictable) identifiers. If that is unavailable, fall back on the xoshiro128+ algorithm to produce random numbers that are not cryptograpically secure, but which do not affect the global random number stream (see Details). If desired, you can produce "predictable" random identifiers that respect the value of the global random number stream via set.seed.

random_id(n = 1, bytes = 16, use_openssl = NULL, global = FALSE)

Arguments

n
number of ids to return. If NULL, it instead returns the generating function

bytes
The number of bytes to include for each identifier. The length of the returned identifiers will be twice this long with each pair of characters representing a single byte.

use_openssl
Optionally a logical, indicating if we should use the openssl for generating the random identifiers from the non-global source. If not given we prefer to use openssl if it is available but fall back on R (See Details).

global
Logical, indicating if random numbers should be global (given R's global random number seed). If TRUE, then ids generated will be predicatable.

Details

Since ids version 1.2.0, the openssl package is optional, and this affects non-global random number drawing. If you have openssl installed your random numbers will be ~50x faster than the implementation we include here.

If global = TRUE we always use a simple link{sample} based algorithm that is driven from the global random number stream. However, when global = FALSE the behaviour depends on the value of use_openssl and whether that package is installed, either using the openssl generators, using an internal algorithm based on xoshiro128+ or erroring.

use_openssl = NULL and openssl installed: openssl

use_openssl = NULL and openssl missing: internal

use_openssl = TRUE and openssl installed: openssl

use_openssl = TRUE and openssl missing: error

use_openssl = FALSE: internal

Author

Rich FitzJohn

Examples

# Generate a random id:
ids::random_id()
#> [1] "deff66240407034f25e7f416bca0795b"

# Generate 10 of them!
ids::random_id(10)
#>  [1] "d38906608dace531cb2681f42204cbde" "657921b26fcf22ce37f1bea0600a5e39"
#>  [3] "ac8c8074c9e45be83e277a2cadcd1cff" "ba2d5ec764e9723069eb26486b78f452"
#>  [5] "171bf6caf9e9b0e53a2138501edefcc3" "17f865eca4cc028d1ab8d2949dcd1590"
#>  [7] "b6108c866ae771f930851b8ffd261925" "a0af3632f687eddbba5d43670ebdc877"
#>  [9] "0a38d345ab53e020c90df61f4322335e" "384867badd5841e275f6737111eb9f03"

# Different length ids
random_id(bytes = 8)
#> [1] "819751feec8e5c6b"
# (note that the number of characters is twice the number of bytes)

# The ids are not affected by R's RNG state:
set.seed(1)
(id1 <- ids::random_id())
#> [1] "c647cb4010c2fd684f26b4d3f4c75e0b"
set.seed(1)
(id2 <- ids::random_id())
#> [1] "b0116d4a085226e1900498eed4d8dc66"
# The generated identifiers are different, despite the seed being the same:
id1 == id2
#> [1] FALSE

# If you need these identifiers to be reproducible, pass use_openssl = FALSE
set.seed(1)
(id1 <- ids::random_id(use_openssl = FALSE))
#> [1] "32d4848f567f69bd51909c6dec72b0bb"
set.seed(1)
(id2 <- ids::random_id(use_openssl = FALSE))
#> [1] "fdee8bad7b6c773dd6555d2f8857e405"
# This time they are the same:
id1 == id2
#> [1] FALSE

# Pass \code{n = NULL} to generate a function that binds your arguments:
id8 <- ids::random_id(NULL, bytes = 8)
id8(10)
#>  [1] "2883ed110dccf09b" "9052e9ac6f60ff4e" "3aef7392f3ec63f5" "5eef95dff155395d"
#>  [5] "141aa50e3a8c6486" "2b0584572e5c3dbc" "816881100bd48e94" "ce08d4f2b86474ac"
#>  [9] "7799a6677e451028" "ee5c75ffff28d703"