Generate UUIDs

Generate UUIDs. In previous versions this was simply a thin wrapper around uuid::UUIDgenerate, however this was subject to collisions on windows where relatively small numbers of UUIDs generated at the same time could return values that were identical. We now generate only version 4 UUIDs (i.e., random though with particular bits set).

uuid(n = 1, drop_hyphens = FALSE, use_time = NA)

Arguments

n
number of ids to return. If NULL, it instead returns the generating function

drop_hyphens
Drop the hyphens from the UUID?

use_time
Unused.

Author

Rich FitzJohn

Examples

# Generate one id
uuid()

#> [1] "4c1a87a6-00e4-42f1-8b23-21efb76a861b"

# Or a bunch
uuid(10)

#>  [1] "4986367d-6049-4c83-a137-5d43287d6647"
#>  [2] "5d2f5193-3e58-4dd0-bec2-5649b6d001f9"
#>  [3] "7620ee11-6fde-4fb9-9684-4e5be0c5318e"
#>  [4] "9e6b32e0-53f2-4784-88d9-3274af7f2540"
#>  [5] "5ff6861a-ae31-416c-8f7f-e24d63b3ff86"
#>  [6] "d76c979f-b9a5-4cb3-aaf8-55bd6588447b"
#>  [7] "f73a6a09-d5c0-4135-80c7-2aef30316170"
#>  [8] "4600df33-917f-41bd-8186-ca747fd7c8cc"
#>  [9] "d0114879-4eaa-4163-8c2a-50a87e8c1c69"
#> [10] "fe23b690-1d4b-4bb8-9d62-5a2fbb7689b7"

# More in the style of random_id()
uuid(drop_hyphens = TRUE)

#> [1] "5b1468c794f34f85b09182fb3313be1f"