ids {ids}	R Documentation

Generic id generating function

Description

Generic id generating function

Usage

ids(n, ..., vals = list(...), style = "snake")

Arguments

n
number of ids to return. If NULL, it instead returns the generating function

...
A number of character vectors

vals
A list of character vectors, instead of ...

style
Style to join words with. Can be one of "Pascal", "camel", "snake", "kebab", "dot", "title", "sentence", "lower", "upper", and "constant".

Value

Either a character vector of length n, or a function of one argument if n is NULL

Author(s)

Rich FitzJohn

Examples

# For an example, please see the vignette

[Package ids version 1.0.1 Index]