Generate UUIDs

Generate UUIDs (Universally Unique IDentifiers). In previous versions this was simply a thin wrapper around uuid::UUIDgenerate, however this was subject to collisions on windows where relatively small numbers of UUIDs generated at the same time could return values that were identical. We now generate only version 4 UUIDs (i.e., random though with particular bits set).

uuid(n = 1, drop_hyphens = FALSE, use_time = NA)

Arguments

n
number of ids to return. If NULL, it instead returns the generating function

drop_hyphens
Drop the hyphens from the UUID?

use_time
Unused.

Author

Rich FitzJohn

Examples

# Generate one id
ids::uuid()
#> [1] "d7c2b795-d411-419c-a28f-90bbdd44ee9c"

# Or a bunch
ids::uuid(10)
#>  [1] "e28fcc20-22fc-4a32-9508-363c88457824"
#>  [2] "3bf2160a-a6c9-4846-a0dd-2351cab1627c"
#>  [3] "ad4a7d84-ef25-4b61-9bc3-015b8083d130"
#>  [4] "e772ccb1-fd70-4cab-945d-178035653885"
#>  [5] "02e6b59a-23de-4373-bc8d-8d550d4d280d"
#>  [6] "f73d4776-2796-4dea-bc6b-9d1996c5131c"
#>  [7] "240734de-9192-4af4-a4cd-ae9e0786c5f1"
#>  [8] "8ee8a8d0-96ba-4a19-86e2-5e9eaa1d97ae"
#>  [9] "af3d1eec-8f71-4704-a3a1-953fecf43750"
#> [10] "7de885d2-d462-4352-adf6-0148b0ab03d8"

# More in the style of random_id()
ids::uuid(drop_hyphens = TRUE)
#> [1] "57b707108b5644b9bec10b819c09c6b1"